+++ requirements.txt
@@ -4,7 +4,7 @@ click>=7.1.1
 click-plugins
 click-spinner
 fiona>=1.8.13.post1
-fsspec[http,s3]
+fsspec
 importlib-metadata
 importlib-resources
 numpy>=1.16
